CRUD skieur complet

+++ SpringFirst/src/main/java/tn/esprit/springfirst/repositories/SkierRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import tn.esprit.springfirst.entities.Skier;
 @Repository
-public interface SkierRepository extends JpaRepository<Skier,Long> {
+public interface SkierRepository extends JpaRepository<Skier,Integer> {
 
 }
